+++ tools/psr/backend/main.go
@@ -7,10 +7,10 @@ import (
 	"fmt"
 	vzlog2 "github.com/verrazzano/verrazzano/pkg/log"
 	vzlog "github.com/verrazzano/verrazzano/pkg/log/vzlog"
-	"github.com/verrazzano/verrazzano/tools/psr/workers/config"
-	"github.com/verrazzano/verrazzano/tools/psr/workers/metrics"
-	"github.com/verrazzano/verrazzano/tools/psr/workers/opensearch"
-	"github.com/verrazzano/verrazzano/tools/psr/workers/spi"
+	"github.com/verrazzano/verrazzano/tools/psr/backend/config"
+	"github.com/verrazzano/verrazzano/tools/psr/backend/metrics"
+	"github.com/verrazzano/verrazzano/tools/psr/backend/spi"
+	"github.com/verrazzano/verrazzano/tools/psr/backend/workers/opensearch"
 	"os"
 	kzap "sigs.k8s.io/controller-runtime/pkg/log/zap"
 )
